chore: fix command call
